# Fix performance in R 4.0.0