--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -45,8 +45,7 @@ Suggests:
     xml2,
     zeallot
 Remotes: 
-    r-lib/pillar,
-    r-lib/rlang#963
+    r-lib/pillar
 VignetteBuilder: 
     knitr
 Encoding: UTF-8