--- a/R/c.R
+++ b/R/c.R
@@ -47,6 +47,6 @@ vec_c <- function(...,
                   .ptype = NULL,
                   .name_spec = NULL,
                   .name_repair = c("minimal", "unique", "check_unique", "universal")) {
-  .External2(vctrs_c, .ptype, .name_spec, .name_repair)
+  .External(vctrs_c, list2(...), .ptype, .name_spec, .name_repair)
 }
 vec_c <- fn_inline_formals(vec_c, ".name_repair")