--- a/R/size.R
+++ b/R/size.R
@@ -83,7 +83,7 @@ vec_size <- function(x) {
 #' @export
 #' @rdname vec_size
 vec_size_common <- function(..., .size = NULL, .absent = 0L) {
-  .External2(vctrs_size_common, .size, .absent)
+  .External(vctrs_size_common, list2(...), .size, .absent)
 }
 
 #' @rdname vec_size