--- a/src/init.c
+++ b/src/init.c
@@ -263,7 +263,7 @@ extern SEXP vctrs_recycle_common(SEXP);
 extern SEXP vctrs_cast_common(SEXP);
 extern SEXP vctrs_rbind(SEXP);
 extern SEXP vctrs_cbind(SEXP);
-extern SEXP vctrs_c(SEXP, SEXP, SEXP, SEXP);
+extern SEXP vctrs_c(SEXP);
 extern SEXP vctrs_new_data_frame(SEXP);
 
 static const R_ExternalMethodDef ExtEntries[] = {
@@ -274,7 +274,7 @@ static const R_ExternalMethodDef ExtEntries[] = {
   {"vctrs_cast_common",                (DL_FUNC) &vctrs_cast_common, 2},
   {"vctrs_rbind",                      (DL_FUNC) &vctrs_rbind, 4},
   {"vctrs_cbind",                      (DL_FUNC) &vctrs_cbind, 4},
-  {"vctrs_c",                          (DL_FUNC) &vctrs_c, 3},
+  {"vctrs_c",                          (DL_FUNC) &vctrs_c, 4},
   {"vctrs_new_data_frame",             (DL_FUNC) &vctrs_new_data_frame, -1},
   {NULL, NULL, 0}
 };