Closes #838 

Requires https://github.com/r-lib/rlang/pull/963 to ensure that the NAMEDNESS of captured `...` is always set to the max value.

Unfortunately R 4.0.0 is going to cause a huge performance regression to `vec_rbind()`. It is very similar to the problem in https://github.com/r-lib/vctrs/issues/984

Here is current master, with R 4.0.0. This is supposed to take ~20ms.

```r
library(vctrs)
library(tibble)

x <- rep(list(data.frame(x = 1)), 10000)

fn <- function(x) {
  vec_rbind(!!!x)
}

# master
bench::mark(fn(x), iterations = 10)
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 fn(x)         285ms    311ms      3.12     764MB     54.2

# this PR
bench::mark(fn(x), iterations = 10)
#> # A tibble: 1 x 6
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 fn(x)        19.8ms   21.7ms      43.0     200KB     18.4
```

The performance regression here is basically the same as in https://github.com/r-lib/vctrs/issues/984, in `df_assign()` we are making a copy of the final output container every time we assign one of the 10000 data frames into it.

The main issue is that `SET_VECTOR_ELT()` now bumps the refcnt of the `value` you are assigning, where it didn't used to do that. This makes `MAYBE_REFERENCED()` return true more often, in cases where we really don't need to make a duplicate. The fix is to switch to `MAYBE_SHARED()`, but also requires a patch to rlang on R < 4.0.0 to ensure that the result of rlang's `rlang_env_dots_list()` C callable returns the captured `...` with max namedness. This is important because it forces us to always copy the dots so we don't accidentally modify by reference (the copying was already happening when we were using `MAYBE_REFERENCED()`).

I've left very detailed notes about this in one of the tests
https://github.com/r-lib/vctrs/pull/1057/files#diff-51c944df297abe15044974191b4f4ebbR593-R627