@lionel- 2020-04-28
I'd rather not rush the ownership fixes in rlang. For reference here are the perf diffs when we change to taking dots with `list2()` at the R level:

```r
bench::mark(
  .External2 = vec_c(1, 1),
  .External = vec_c2(1, 1),
  .Extlist2 = vec_c3(1, 1),
  list2 = list2(1, 1),
  base = c(1, 1),
  iterations = 100000,
  check = FALSE
)[1:8]
#> # A tibble: 5 x 8
#>   expression      min   median `itr/sec` mem_alloc `gc/sec`  n_itr  n_gc
#>   <bch:expr> <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>  <int> <dbl>
#> 1 .External2   3.69µs   4.72µs   198184.        0B     5.95  99997     3
#> 2 .External    4.09µs   5.17µs   180357.        0B     5.41  99997     3
#> 3 .Extlist2    5.19µs   6.38µs   146162.        0B     7.31  99995     5
#> 4 list2        1.73µs   2.43µs   381401.        0B    11.4   99997     3
#> 5 base          163ns    193ns  4470507.        0B     0    100000     0
```

In the second line we change `.External2()` by `.External()`. In the third line we use `.External()` with dots captured at the R level. The fourth and fifth lines show `list2()` and `c()` for reference. In any case we're still much much slower than `c()`, but if possible I'd prefer to keep collecting dots at C level. Also it's probably worth keeping `.External2()` for performance-critical functions.

That said, this isn't very important for dplyr 1.0. It will become more important once we export `v()` from funs and people start using `vec_c()` in their scripts, including in loops. In that case we'll be able to improve performance by:

* Collecting dots faster in simple cases: r-lib/rlang#937.
* Bypassing all the coercion stuff for non-S3 base types.

Since that's not important right now, and since the performance difference with `c()` is still so big, I suggest we just collect dots at the R level, which should fix the refcounts.