@DavisVaughan 2020-04-28
@lionel- ready for another round of review. Dropped the rlang Remote in favor of using `list2()` and `dots_values()`, as you suggested.