@lionel- 2020-04-28
Could you add other more specific tests to make sure the refcounts are correct and we're not mutating inputs please? This repair-names test seems a bit indirect and it might be refactored away in the future.